Check if opts.exclude is an array and if not wrap the value in one.
This was noticed when using the CLI and passing only one dependency to exclude. Minimist sees this as only a string and does not provide any method for casting to an array so the value that is passed to wiredep for the exclude parameter ends up to be a string and this messes up the filterExcludedDependencies method.