fix issue #107: throw error when component is not found
see issue #107.
see pull request #137.

Folks can run into issue #107 when they are missing some of the dependencies
that are supposed to be installed with `bower install`. The trick there
is that **some** of the dependencies are installed, otherwise it would fail
with a different error (e.g. cannot find `bower_components`). I can reproduce
the error consistently by cloning a repo at a previous revision, running
`bower install` then doing a `git pull` which changes the list of dependencies
that should be installed and running `grunt wiredep`.

If you are part of a big organization where many folks could be updating
bower dependencies and folks don't run `bower install` religiously after a
`git pull`, then you could run into issue #107 more often.

The difference between this pull request and pull request #137 is that
the latter one allows to configure the error handler when a component is not
found; which may be useful in situations where `wiredep` is called by higher
level tools and you want e.g. to customize the error message.